+++ src/Manager.php
